requirements.txt
# 电子相册Skill依赖包

# 核心依赖
Pillow>=10.0.0          # 图像处理
img2pdf>=0.4.4          # PDF生成

# 可选依赖（飞书文档功能需要）
# feishu-python-sdk>=1.0.0  # 飞书API（如果需要飞书文档功能）

# 开发依赖
pytest>=7.0.0           # 测试框架
black>=23.0.0           # 代码格式化
flake8>=6.0.0           # 代码检查